updated to easymock 3. bumped up version to 2.0.6
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@270 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ test/org/jcvi/assembly/contig/qual/TestAbstractQualityValueStrategyWithComplimentedRead.java
@@ -30,9 +30,9 @@ import org.jcvi.sequence.SequenceDirection;
 public class TestAbstractQualityValueStrategyWithComplimentedRead extends TestAbstractQualityValueStrategy{
     
     private int convertToFullRangeComplimentedIndex(int index){
-        int switchedValidLeft = LENGTH - (int)validRange.getEnd();
-        int distanceFromLeftValidRange = index+switchedValidLeft;
-        return LENGTH - distanceFromLeftValidRange;
+        long switchedValidLeft = LENGTH - validRange.getEnd();
+        long distanceFromLeftValidRange = index+switchedValidLeft;
+        return (int)(LENGTH - distanceFromLeftValidRange);
     }
 
     @Override
